+++ src/router/index.js
@@ -1,6 +1,5 @@
 import { createWebHistory, createRouter } from 'vue-router';
 import AIFeedback from '../components/pages/AIFeedback.vue';
-import AIFeedback_2 from '../components/pages/AIFeedback_2.vue';
 
 const routes = [
     {
@@ -10,10 +9,6 @@ const routes = [
             {
                 path: '/feedback/ai',
                 component: AIFeedback,
-            },
-            {
-                path: '/feedback/ai2',
-                component: AIFeedback_2,
             },
         ],
     },
